Move JSON Schema node helpers off the strategy into JsonSchemaNodes
applyDescription is a stateless utility and does not belong on the converter
strategy. Introduce a package-private JsonSchemaNodes helper for the cohesive
set of JSON Schema node helpers, and fold in $ref-node construction that was
duplicated across the reference and union converters.
