+++ kinotic-idl/src/main/java/org/kinotic/idl/api/converter/jsonschema/McpJsonSchemaGenerator.java
@@ -47,7 +47,7 @@ public class McpJsonSchemaGenerator {
 
         for (ParameterDefinition parameter : function.getParameters()) {
             ObjectNode parameterSchema = converter.convert(parameter.getType());
-            JsonSchemaConverterStrategy.applyDescription(parameterSchema, parameter);
+            JsonSchemaNodes.applyDescription(parameterSchema, parameter);
             properties.set(parameter.getName(), parameterSchema);
 
             if (parameter.containsDecorator(NotNullC3Decorator.class)) {
